% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.r
\name{daylength}
\alias{daylength}
\title{Calculates day length (in hours) and the solar elevation}
\usage{
daylength(doy, latitude)
}
\arguments{
\item{doy}{a vector with doy values 1 - 365(6)}

\item{latitude}{a given latitude}
}
\value{
nested list with daylength (daylength) and
solar elevation (solar_elev) elements
}
\description{
This routine uses Forsythe et al. 1995.
}
\examples{

\donttest{
# calcualte the hours of sunlight and solar elevation on day of year 1
# and latitude 51
ephem <- daylength(1, 51)
print(ephem)
}
}
\keyword{ephemerids}
\keyword{solar,}
