% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.outliers.r
\name{detect.outliers}
\alias{detect.outliers}
\title{Detect outliers in PhenoCam time series}
\usage{
detect.outliers(data, iterations = 10, vis = FALSE)
}
\arguments{
\item{data:}{PhenoCam data frame or filename}

\item{iterations:}{number of itterations over which to cycle to detect outliers ()}
}
\description{
Detect outliers in PhenoCam time series
}
\examples{
# with defaults, overwriting the original data frame:
df <- detect.outliers(df, iterations=5)
}
\keyword{PhenoCam,}
\keyword{outliers,}
\keyword{pre-processing}
\keyword{smoothing,}

