% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.r
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers in PhenoCam time series (data frame or file)
The function fills in the existing column to hold outlier flags,
and either overwrites the original file or outputs a new data frame.}
\usage{
detect_outliers(data, iterations = 20, sigma = 2, grvi = FALSE,
  snowflag = FALSE, plot = FALSE)
}
\arguments{
\item{data:}{PhenoCam data frame or filename}

\item{iterations:}{number of itterations over which to cycle to detect outliers ()}

\item{sigma:}{number of deviations to exclude outliers at}

\item{grvi:}{reverse the direction of the screening intervals
to accomodate for GRVI outliers}

\item{plot:}{visualize the process, mostly for debugging (TRUE / FALSE)}

\item{snowflag:}{integrate snow flags?}
}
\description{
Detect outliers in PhenoCam time series (data frame or file)
The function fills in the existing column to hold outlier flags,
and either overwrites the original file or outputs a new data frame.
}
\examples{

\dontrun{
# download demo data (do not detect outliers)
download.phenocam(site = "harvard",
                  vegetation = "DB",
                  roi_id = 1,
                  frequency = 3,
                  outlier_detection = FALSE)

# detect outliers in the downloaded file
detect_outliers("harvard_DB_0001_1day.csv")

# This function is rarely used stand-alone as it is called by the
# download_phenocam() function and does not serve any other purpose
# as it needs the outlier_flag_xxx field to work (and does not verify this).
# Use stand-alone at your own risk.
}
}
\keyword{PhenoCam,}
\keyword{outliers,}
\keyword{pre-processing}
\keyword{smoothing,}
