% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.phenocam.r
\name{download.phenocam}
\alias{download.phenocam}
\title{Function to download PhenoCam time series based upon:
site name, vegetation type, geographic bounding box.
You can optionally merge them with Daymet climate data
when located within the Daymet data range.}
\usage{
download.phenocam(site = "bartlett", vegetation = "all", frequency = "3",
  roi_id = NULL, top_left = NULL, bottom_right = NULL, daymet = FALSE,
  trim_daymet = TRUE, outlier_detection = FALSE, phenophase = FALSE,
  out_dir = getwd())
}
\arguments{
\item{site}{: the site name, as mentioned on the PhenoCam web page.}

\item{vegetation}{: vegeation type (DB, EN, ... default = ALL)}

\item{frequency}{: frequency of the PhenoCam time series product, in days (1,3, "raw" are allowed)}

\item{daymet}{: TRUE or FALSE, if true merges the daymet data at the location (only for frequency 1 and 3)}

\item{roi_id:}{the id of the ROI to download, if not specified download all}

\item{top_lef:}{latitude, longitude tupple as a vector or a list (c(45.5, -80)), requires bottom_right}

\item{bottom_right:}{latitude, longitude tupple as a vector or a list (c(45.5, -80)), requires top_left}

\item{trim_daymet:}{TRUE or FALSE, if TRUE only download Daymet data for the dates matching the Gcc data. If FALSE, provide the whole Daymet series (for modelling spinup purposes).}

\item{outlier_detection:}{TRUE or FALSE, detect outliers}

\item{phenophase:}{TRUE or FALSE, calculate threshold based phenological transition dates (this will generate a non standard product)}
}
\description{
Function to download PhenoCam time series based upon:
site name, vegetation type, geographic bounding box.
You can optionally merge them with Daymet climate data
when located within the Daymet data range.
}
\examples{
download.phenocam(site="harvard",vegetation="DB",frequency=3,daymet=TRUE)
}
\keyword{DAYMET,}
\keyword{PhenoCam,}
\keyword{climate}
\keyword{data,}
\keyword{modelling}

