% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_phenocam.r
\name{explore_phenocam}
\alias{explore_phenocam}
\title{Starts the phenocamr shiny interface
in your default browser}
\usage{
explore_phenocam()
}
\arguments{
\item{none}{: calling the function loads the GUI}
}
\description{
Starts the phenocamr shiny interface
in your default browser
}
\examples{

\dontrun{
# Starts the PhenoCam explorer in a browser
explore_phenocam()
}
}
\keyword{GUI,}
\keyword{end,}
\keyword{front}
\keyword{interactive}
