% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grvi.r
\name{grvi}
\alias{grvi}
\title{Calculate GRVI for a PhenoCam file or data frame}
\usage{
grvi(df, par = c(1, 1, 1))
}
\arguments{
\item{df}{a PhenoCam data file or data frame (when using a file provide a
full path if not in the current working directory)}

\item{par}{grvi parameters (weights)}
}
\description{
Calculate GRVI for a PhenoCam file or data frame
}
\examples{

\donttest{
# with defaults, outputting a data frame
# with smoothed values, overwriting the original

# download demo data
download_phenocam(site = "harvard$",
                  vegetation = "DB",
                  roi_id = "1",
                  frequency = "3")

# calculate and append the GRVI for a file (overwrites the original)
grvi(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))

# the function also works on a PhenoCam data frame
# but you will lose the extensive header meta-data in the process
df = utils::read.csv(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))
df = grvi(df)
}
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
