% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_rois.r
\name{list_rois}
\alias{list_rois}
\title{Function to list all site regions-of-interst (ROIs)}
\usage{
list_rois(out_dir = tempdir(), internal = TRUE)
}
\arguments{
\item{out_dir}{output directory (default = tempdir())}

\item{internal}{TRUE or FALSE (default = TRUE)}
}
\description{
Function to list all site regions-of-interst (ROIs)
}
\examples{

\donttest{
# download the site meta-data
df <- list_rois()
}
}
\keyword{PhenoCam,}
\keyword{meta-data}
