% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_daymet.r
\name{merge_daymet}
\alias{merge_daymet}
\title{Merge Daymet data with a PhenoCam data file.}
\usage{
merge_daymet(filename, trim_daymet = FALSE)
}
\arguments{
\item{filename}{a PhenoCam data file}

\item{trim_daymet}{logical, trim the daymet data to the length of the
PhenoCam time series or include the whole Daymet time series (1980-current).
(default = \code{FALSE})}
}
\description{
Merge Daymet data with a PhenoCam data file.
}
\examples{

\donttest{
# download demo data (do not smooth)
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1",
                  frequency = "3")

# merge data with daymet data
merge_daymet(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))
}
}
\keyword{Daymet,}
\keyword{PhenoCam,}
\keyword{data}
\keyword{integration}
\keyword{series,}
\keyword{time}
