% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_modis.r
\name{merge_modis}
\alias{merge_modis}
\title{Merge MODIS data with a PhenoCam data file.
Merged data includes MODIS MCD43 BRDF data for
bands 1 - 3 and matching QA/QC data as well as
MODIS phenology (MCD12Q2) data}
\usage{
merge_modis(df, size = c(0, 0))
}
\arguments{
\item{df:}{a PhenoCam data file (not a data frame!)}
}
\description{
Merge MODIS data with a PhenoCam data file.
Merged data includes MODIS MCD43 BRDF data for
bands 1 - 3 and matching QA/QC data as well as
MODIS phenology (MCD12Q2) data
}
\examples{

\dontrun{
# download demo data (do not smooth)
download_phenocam(site = "harvard",
                  vegetation = "DB",
                  roi_id = 1,
                  frequency = 3)
merge_modis("harvard_DB_0001_1day.csv")
}
}
\keyword{MODIS,}
\keyword{PhenoCam,}
\keyword{data}
\keyword{integration}
\keyword{series,}
\keyword{time}
