% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.ts.r
\name{normalize.ts}
\alias{normalize.ts}
\title{Normalize PhenoCam time series, between 0-1 to to standardize
further processing, independent of the relative amplitude
of the time series (works on vectors not data frames)}
\usage{
normalize.ts(df, percentile = 90)
}
\arguments{
\item{df:}{a PhenoCam data frame}
}
\description{
Normalize PhenoCam time series, between 0-1 to to standardize
further processing, independent of the relative amplitude
of the time series (works on vectors not data frames)
}
\examples{

# Internal function only, should not be used stand-alone.
# As such no documentation is provided.
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}

