% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_ts.r
\name{normalize_ts}
\alias{normalize_ts}
\title{Normalize PhenoCam time series}
\usage{
normalize_ts(df, percentile = 90)
}
\arguments{
\item{df}{a PhenoCam data frame}

\item{percentile}{percentile value to interprete}
}
\description{
Normalize PhenoCam data between 0-1 to to standardize
further processing, independent of the relative amplitude
of the time series (works on vectors not data frames)
}
\examples{

# Internal function only, should not be used stand-alone.
# As such no documentation is provided.
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
