% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.span.r
\name{optimal.span}
\alias{optimal.span}
\title{Calculates the optimal span for a loess spline
smoother based upon the bayesian information
criterion (BCI).}
\usage{
optimal.span(y, x = NULL, weights = NULL, step = 0.005)
}
\arguments{
\item{y:}{a vector with measurement values to smooth}

\item{x:}{a vector with dates / time steps}

\item{weights:}{optional values to weigh the loess fit with}

\item{step:}{span increment size}
}
\description{
Calculates the optimal span for a loess spline
smoother based upon the bayesian information
criterion (BCI).
}
\examples{
span <- optimal.span(y)

# dates need to be converted to standard notation using
as.Date(my_dates)
}
\keyword{loess,}
\keyword{series}
\keyword{smoother,}
\keyword{span,}
\keyword{time}

