% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenocam_explorer.r
\name{phenocam_explorer}
\alias{phenocam_explorer}
\title{Starts the phenocamr shiny interface
in your default browser. The GUI
allows you to interactively download
data and visualize timeseries.}
\usage{
phenocam_explorer()
}
\description{
Starts the phenocamr shiny interface
in your default browser. The GUI
allows you to interactively download
data and visualize timeseries.
}
\examples{

\donttest{
# Starts the PhenoCam explorer GUI in a browser
phenocam_explorer()
}
}
\keyword{GUI,}
\keyword{end,}
\keyword{front}
\keyword{interactive}
