% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenophases.r
\name{phenophases}
\alias{phenophases}
\title{Calculates phenophases, both greenup and senescence, for all 'seasons'
in a PhenoCam time series. This routine combines a forward and backward
run of transition_dates function to calculate the phenophases.}
\usage{
phenophases(data, mat = NULL, internal = TRUE, out_dir = tempdir(), ...)
}
\arguments{
\item{data}{a PhenoCam data file (or data frame)}

\item{mat}{mean annual temperature}

\item{internal}{return PhenoCam data file or data frame}

\item{out_dir}{output directory}

\item{...}{pass parameters to the transition_dates() function}
}
\description{
Calculates phenophases, both greenup and senescence, for all 'seasons'
in a PhenoCam time series. This routine combines a forward and backward
run of transition_dates function to calculate the phenophases.
}
\examples{

\donttest{
# downloads a time series for Bartlett Forest data and calculates
# the matching phenophases.
# Outputs a nested list of phenophases dates
# where location [[1]] holds the greenup dates and location
# [[2]] the greendown dates
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1",
                  frequency = "3")
                  
df = utils::read.csv(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))
my_dates = phenophases(df, output = FALSE)
}
}
\keyword{PhenoCam,}
\keyword{dates,}
\keyword{phenology,}
\keyword{series}
\keyword{time}
\keyword{transition}
