% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phenocam.r
\name{process_phenocam}
\alias{process_phenocam}
\title{Function to post-process PhenoCam time series}
\usage{
process_phenocam(file, outlier_detection = TRUE, smooth = TRUE,
  contract = FALSE, daymet = FALSE, trim_daymet = TRUE,
  trim = NULL, phenophase = FALSE, out_dir = tempdir(),
  internal = FALSE, snow_flag = FALSE, snow_weight = 1,
  save_gcc90 = FALSE, weights_as_n = FALSE, span = NULL)
}
\arguments{
\item{file}{1 or 3-day PhenoCam time series file path}

\item{outlier_detection}{TRUE or FALSE, detect outliers}

\item{smooth}{smooth data (logical, default is \code{TRUE})}

\item{contract}{contract 3-day data (logical, default is \code{TRUE})}

\item{daymet}{TRUE or FALSE, merges the daymet data}

\item{trim_daymet}{TRUE or FALSE, trims data to match PhenoCam data}

\item{trim}{year (numeric) to which to constrain the output (default = \code{NULL})}

\item{phenophase}{logical, calculate transition dates (default = \code{FALSE})}

\item{out_dir}{output directory where to store downloaded data 
(default = tempdir())}

\item{internal}{allow for the data element to be returned to the workspace}

\item{snow_flag}{integrate snow flags?}

\item{snow_weight}{weight given to snow flagged values, by default these are
not down weighted in smoothing (default = 1)}

\item{save_gcc90}{whether to save the plot output for gcc_90 timeseries (as a test)}

\item{weights_as_n}{whether to use sum of the weights in AIC formula}

\item{span}{customized span list or variable, default is NULL}
}
\value{
Downloaded files in out_dir of requested time series products, as well
as derived phenophase estimates based upon these time series.
}
\description{
Wrapper around other more basic funtions
}
\examples{

\donttest{
# download the first ROI time series for the Harvard PhenoCam site
# at an aggregation frequency of 3-days.
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1000",
                  frequency = "3")
 
# read phenocam data into phenocamr data structure                  
df <- process_phenocam(file.path(tempdir(),"harvard_DB_1000_3day.csv"))
                  
}
}
\keyword{Daymet,}
\keyword{PhenoCam,}
\keyword{climate}
\keyword{data,}
\keyword{modelling,}
\keyword{post-processing}
