% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.ts.r
\name{smooth.ts}
\alias{smooth.ts}
\title{Smooths a standard PhenoCam file or data frame using
the BCI optimized smoothing parameter}
\usage{
smooth.ts(df, force = TRUE)
}
\arguments{
\item{df:}{a PhenoCam data file or data frame}
}
\description{
Smooths a standard PhenoCam file or data frame using
the BCI optimized smoothing parameter
}
\examples{

# with defaults, outputting a data frame
# with smoothed values, overwriting the original

# download demo data (do not smooth)
# download.phenocam(site="harvard",
#                   vegetation="DB",
#                  roi_id=1,
#                   frequency=3,
#                  smooth=FALSE)

# smooth the downloaded file (and overwrite the original)
# smooth.ts("harvard_DB_0001_1day.csv")

# the function also works on a PhenoCam data frame
# but you will lose the extensive header in the process
# df = read.csv("harvard_DB_0001_1day.csv")
# df = smooth.ts(df)
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}

