% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.flag.r
\name{snow.flag}
\alias{snow.flag}
\title{Flags snowy pictures based upon Citizen Science analysis of the imagery (zooniverse)}
\usage{
snow.flag(filename)
}
\arguments{
\item{filename:}{filename of a downloaded time series}
}
\description{
Flags snowy pictures based upon Citizen Science analysis of the imagery (zooniverse)
}
\keyword{PhenoCam,}
\keyword{outliers}
\keyword{pre-processing,}
\keyword{screening,}
\keyword{snow,}

