% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_flag.r
\name{snow_flag}
\alias{snow_flag}
\title{Flags snowy pictures based upon Citizen Science analysis
of the select imagery (zooniverse). Will be included in dataset but
currently does not serve any real purpose.}
\usage{
snow_flag(df)
}
\arguments{
\item{df}{a PhenoCam data file or data frame}
}
\description{
Flags snowy pictures based upon Citizen Science analysis
of the select imagery (zooniverse). Will be included in dataset but
currently does not serve any real purpose.
}
\examples{

# Internal function only, should not be used stand-alone.
# As such no documentation is provided.
}
\keyword{PhenoCam,}
\keyword{outliers}
\keyword{pre-processing,}
\keyword{screening,}
\keyword{snow,}
