% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract_phenocam.r
\name{contract_phenocam}
\alias{contract_phenocam}
\title{Contracts the file from 1-day to a 3-day time step
reverting the expand.phenocam() routine in order
to save space and generate files as outlined in
the data paper. This routine is used as a post-production
measure.}
\usage{
contract_phenocam(filename)
}
\arguments{
\item{filename}{a phenocam data file with a 3 day time step}
}
\description{
Contracts the file from 1-day to a 3-day time step
reverting the expand.phenocam() routine in order
to save space and generate files as outlined in
the data paper. This routine is used as a post-production
measure.
}
\examples{

\donttest{
# download demo data
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1",
                  frequency = "3")

# Overwrites the original file, increasing
# it's file size.
expand_phenocam(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))

# Contracts the file to it's original size, skipping
# two days.
contract_phenocam(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))
}
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
