% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract_phenocam.r
\name{contract_phenocam}
\alias{contract_phenocam}
\title{Contracts the file from 1-day to a 3-day time step
reverting the expand.phenocam() routine in order
to save space and generate files as outlined in
the data paper. This routine is used as a post-production
measure, and not used in general manipulation
of the data as even 1-day files are sufficiently
small. Once expanded they are kept this way.}
\usage{
contract_phenocam(filename)
}
\arguments{
\item{file:}{a phenocam data file with a 3 day time step}
}
\description{
Contracts the file from 1-day to a 3-day time step
reverting the expand.phenocam() routine in order
to save space and generate files as outlined in
the data paper. This routine is used as a post-production
measure, and not used in general manipulation
of the data as even 1-day files are sufficiently
small. Once expanded they are kept this way.
}
\examples{
# df = download.phenocam(site = "harvard",
#                        type = "DB",
#                        roi = "1",
#                        frequency = 3)

# expand.phenocam("harvard_DB_0001_1day.csv")
# Overwrites the original file, increasing
# it's file size.

# contract_phenocam("harvard_DB_0001_1day.csv")
# Contracts the file to it's original size, skipping
# two days.
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
