% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.phenocam.r
\name{download.phenocam}
\alias{download.phenocam}
\title{Function to download PhenoCam time series based upon:
site name, vegetation type, geographic bounding box.
You can optionally merge them with Daymet climate data
when located within the Daymet data range.}
\usage{
download.phenocam(site = "bartlett", vegetation = NULL, frequency = "3",
  roi_id = NULL, top_left = NULL, bottom_right = NULL,
  outlier_detection = TRUE, smooth = TRUE, daymet = FALSE,
  trim_daymet = TRUE, phenophase = FALSE, out_dir = getwd())
}
\arguments{
\item{site}{: the site name, as mentioned on the PhenoCam web page.}

\item{vegetation}{: vegeation type (DB, EN, ... default = ALL)}

\item{frequency}{: frequency of the PhenoCam time series product, in days (1,3, "raw" are allowed)}

\item{daymet}{: TRUE or FALSE, if true merges the daymet data at the location (only for frequency 1 and 3)}

\item{roi_id:}{the id of the ROI to download, if not specified download all}

\item{top_lef:}{latitude, longitude tupple as a vector c(45.5, -80), requires bottom_right}

\item{bottom_right:}{latitude, longitude tupple as a vector c(45.5, -80), requires top_left}

\item{smooth:}{smooth data for future plotting (TRUE / FALSE, default is TRUE)}

\item{trim_daymet:}{TRUE or FALSE, if TRUE only download Daymet data for the dates matching the Gcc data. 
If FALSE, I provide the whole Daymet series (for modelling spin-up purposes).}

\item{outlier_detection:}{TRUE or FALSE, detect outliers}

\item{phenophase:}{TRUE or FALSE, calculate threshold based phenological transition dates (this will generate a non standard product)}
}
\description{
Function to download PhenoCam time series based upon:
site name, vegetation type, geographic bounding box.
You can optionally merge them with Daymet climate data
when located within the Daymet data range.
}
\examples{

# download the first ROI time series for the Harvard PhenoCam site
# and an aggregation frequency of 3-days.
# download.phenocam(site="harvard",
#                   vegetation="DB",
#                   roi_id=1,
#                   frequency=3)
                  
# download all Harvard Forest deciduous broadleaf sites using
# geographic constraints.
# download.phenocam(vegetation="DB",
#                   roi_id=1,
#                   frequency=3,
#                   top_left=c(42.545657, -72.197524),
#                   bottom_right=c(42.527079, -72.158128))
}
\keyword{Daymet,}
\keyword{PhenoCam,}
\keyword{climate}
\keyword{data,}
\keyword{modelling}

