% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_phenocam.r
\name{expand_phenocam}
\alias{expand_phenocam}
\title{Expand a PhenoCam file from 3-day to a 1-day time step
All values will be filled with NA. This is usefull
to store subsequent interpolated (1-day data)
and increase consistency across processing.}
\usage{
expand_phenocam(df, truncate = NULL)
}
\arguments{
\item{df}{a PhenoCam data frame}

\item{truncate}{year (numerical), limit the time series
to a particular year}
}
\description{
Expand a PhenoCam file from 3-day to a 1-day time step
All values will be filled with NA. This is usefull
to store subsequent interpolated (1-day data)
and increase consistency across processing.
}
\examples{

\dontrun{
df = download.phenocam(site = "harvard",
                       type = "DB",
                       roi = "1",
                       frequency = 3)

# expand.phenocam("harvard_DB_0001_1day.csv")
# overwrites the original file, increasing
# it's file size. shrink the file using
contract.phenocam()
}
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
