% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_phenocam.r
\name{expand_phenocam}
\alias{expand_phenocam}
\title{Expand a PhenoCam file from 3-day to a 1-day time step
All values will be filled with NA. This is usefull
to store subsequent interpolated (1-day data)
and increase consistency across processing.}
\usage{
expand_phenocam(df, truncate = NULL)
}
\arguments{
\item{df}{a PhenoCam data frame}

\item{truncate}{year (numerical), limit the time series
to a particular year}
}
\description{
Expand a PhenoCam file from 3-day to a 1-day time step
All values will be filled with NA. This is usefull
to store subsequent interpolated (1-day data)
and increase consistency across processing.
}
\examples{

\donttest{
# download demo data
download_phenocam(site = "harvard$",
                  vegetation = "DB",
                  roi_id = "1",
                  frequency = "3")

# Overwrites the original file, increasing
# it's file size.
expand_phenocam(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))

# Contracts the file to it's original size, skipping
# two days.
contract_phenocam(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))
}
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
