% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_phenocam.r
\name{explore_phenocam}
\alias{explore_phenocam}
\title{Starts the phenocamr shiny interface
in your default browser. The GUI
allows you to interactively download
data and visualize timeseries.}
\usage{
explore_phenocam()
}
\description{
Starts the phenocamr shiny interface
in your default browser. The GUI
allows you to interactively download
data and visualize timeseries.
}
\examples{

\dontrun{
# Starts the PhenoCam explorer GUI in a browser
explore_phenocam()
}
}
\keyword{GUI,}
\keyword{end,}
\keyword{front}
\keyword{interactive}
