% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grvi.r
\name{grvi}
\alias{grvi}
\title{Calculate GRVI for a PhenoCam file or data frame}
\usage{
grvi(df, par = c(1, 1, 1), plot = FALSE)
}
\arguments{
\item{df}{a PhenoCam data file or data frame}

\item{par}{grvi parameters (weights)}

\item{plot}{plot the data while being evaluated}
}
\description{
Calculate GRVI for a PhenoCam file or data frame
}
\examples{

\dontrun{
# with defaults, outputting a data frame
# with smoothed values, overwriting the original

# download demo data (do not smooth)
download_phenocam(site = "harvard",
                  vegetation = "DB",
                  roi_id = 1,
                  frequency = 3,
                  smooth = FALSE)

# calculate and append the GRVI for a file (and overwrite the original)
grvi("harvard_DB_0001_1day.csv")

# the function also works on a PhenoCam data frame
# but you will lose the extensive header in the process
df = read.csv("harvard_DB_0001_1day.csv")
df = grvi(df)
}
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
