% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grvi.r
\name{grvi}
\alias{grvi}
\title{Calculate green-red vegetation index (GRVI) for a PhenoCam file or data frame}
\usage{
grvi(df, par = c(1, 1, 1))
}
\arguments{
\item{df}{a PhenoCam data file or data frame (when using a file provide a
full path if not in the current working directory)}

\item{par}{grvi parameters (digital number weights)}
}
\description{
The GRVI is defined as the normalized ratio between the red and green channel
of a RGB image or digital number triplet. However, the blue channel can be
used as well using a weighting factor. As such a paramter vector is provided
so different channels / DN can be weighted separately.
}
\examples{

\donttest{
# with defaults, outputting a data frame
# with smoothed values, overwriting the original

# download demo data
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1",
                  frequency = "3")

# calculate and append the GRVI for a file (overwrites the original)
grvi(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))

# the function also works on a PhenoCam data frame
# but you will lose the extensive header meta-data in the process
df = utils::read.csv(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))
df = grvi(df, par = c(1, 1, 0))
}
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
