% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_rois.r
\name{list_rois}
\alias{list_rois}
\title{List all site regions-of-interst (ROIs)}
\usage{
list_rois(out_dir = tempdir(), internal = TRUE)
}
\arguments{
\item{out_dir}{output directory (default = tempdir())}

\item{internal}{TRUE or FALSE (default = TRUE)}
}
\value{
A data frame with ROIs for all available cameras
}
\description{
The ROI list can be helpful in determining which time series to download
using `download_phenocam()`.
}
\examples{

\dontrun{
# download the site meta-data
df <- list_rois()
}
}
