% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sites.r
\name{list_sites}
\alias{list_sites}
\title{Function to list site meta-data}
\usage{
list_sites(out_dir = tempdir(), internal = TRUE)
}
\arguments{
\item{out_dir}{output directory (default = tempdir())}

\item{internal}{TRUE or FALSE (default = TRUE)}
}
\description{
Function to list site meta-data
}
\examples{

\dontrun{
# download the site meta-data
df <- list_sites()
}
}
\keyword{PhenoCam,}
\keyword{meta-data}
