% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_daymet.r
\name{merge_daymet}
\alias{merge_daymet}
\title{Merge Daymet data with a PhenoCam data file.}
\usage{
merge_daymet(df, trim_daymet = FALSE)
}
\arguments{
\item{df}{a PhenoCam data file}

\item{trim_daymet}{TRUE / FALSE, trim the daymet data to the length of the
PhenoCam time series or include the whole Daymet time series (1980-current).}
}
\description{
Merge Daymet data with a PhenoCam data file.
}
\examples{

\dontrun{
# download demo data (do not smooth)
download_phenocam(site = "harvard",
                  vegetation = "DB",
                  roi_id = 1,
                  frequency = 3)
merge_daymet("harvard_DB_0001_1day.csv")
}
}
\keyword{Daymet,}
\keyword{PhenoCam,}
\keyword{data}
\keyword{integration}
\keyword{series,}
\keyword{time}
