% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.ts.r
\name{normalize.ts}
\alias{normalize.ts}
\title{Normalize time series, between 0-1 to to standardize
further processing, independent of the relative amplitude
of the time series}
\usage{
normalize.ts(df)
}
\arguments{
\item{df:}{a PhenoCam data file or data frame}
}
\description{
Normalize time series, between 0-1 to to standardize
further processing, independent of the relative amplitude
of the time series
}
\examples{
# with defaults, outputting a data frame
# with normalized values (0-1), overwriting the original
df <- normalize.ts(df)
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}

