% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenophases.r
\name{phenophases}
\alias{phenophases}
\title{Calculates phenophases for all 'seasons' in a PhenoCam
time series. This routine combines a forward and backward
run of transition.dates() to calculate all phenophases.}
\usage{
phenophases(df, plot = TRUE, pdf = FALSE)
}
\arguments{
\item{df:}{a PhenoCam data file or data frame}

\item{plot:}{provide a basic graph of the data and estimated
dates}

\item{pdf:}{output plot as pdf in current working directory}
}
\description{
Calculates phenophases for all 'seasons' in a PhenoCam
time series. This routine combines a forward and backward
run of transition.dates() to calculate all phenophases.
}
\examples{
# with defaults, outputting a nested list of phenophases dates
# where location [[1]] holds the greenup dates and location
[[2]] the greendown dates

my_dates <- phenophases(df,plot=TRUE)

# dates need to be converted to standard notation using
as.Date(my_dates)
}
\keyword{PhenoCam,}
\keyword{dates,}
\keyword{phenology,}
\keyword{series}
\keyword{time}
\keyword{transition}

