% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenocam.r
\name{read_phenocam}
\alias{read_phenocam}
\title{Read phenocam time series data}
\usage{
read_phenocam(filename)
}
\arguments{
\item{filename}{a PhenoCam data file}
}
\description{
Data is read into a nested list, preserving
header data and critical file name information.
}
\examples{

\donttest{
# with defaults, outputting a data frame
# with smoothed values, overwriting the original

# download demo data (do not smooth)
download_phenocam(site = "harvard$",
                  veg_type = "DB",
                  roi_id = "1",
                  frequency = "3",
                  smooth = FALSE)

# read the phenocamo data file
df = read_phenocam(paste0(tempdir(),"/harvard_DB_0001_3day.csv"))

# print data structure
print(str(df))

# write the phenocamo data file
write_phenocam(df, out_dir = tempdir())
}
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
