% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.ts.r
\name{smooth.ts}
\alias{smooth.ts}
\title{Smooths a standard PhenoCam file or data frame using
the BCI optimized smoothing parameter}
\usage{
smooth.ts(df)
}
\arguments{
\item{df:}{a PhenoCam data file or data frame}
}
\description{
Smooths a standard PhenoCam file or data frame using
the BCI optimized smoothing parameter
}
\examples{
# with defaults, outputting a data frame
# with smoothed values, overwriting the original
df <- smooth.ts(df)
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}

