% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_ts.r
\name{smooth_ts}
\alias{smooth_ts}
\title{Smooths a standard PhenoCam file or data frame using
the BCI optimized smoothing parameter}
\usage{
smooth_ts(df, metrics = c("gcc_mean", "gcc_50", "gcc_75", "gcc_90",
  "rcc_mean", "rcc_50", "rcc_75", "rcc_90"), force = TRUE)
}
\arguments{
\item{df:}{a PhenoCam data file or data frame}

\item{metrics:}{which metrics to process, normally all default ones}

\item{force:}{TRUE / FALSE, force reprocessing?}
}
\description{
Smooths a standard PhenoCam file or data frame using
the BCI optimized smoothing parameter
}
\examples{

\dontrun{
# with defaults, outputting a data frame
# with smoothed values, overwriting the original

# download demo data (do not smooth)
download.phenocam(site="harvard",
                  vegetation="DB",
                  roi_id=1,
                  frequency=3,
                  smooth=FALSE)

# smooth the downloaded file (and overwrite the original)
smooth_ts("harvard_DB_0001_1day.csv")

# the function also works on a PhenoCam data frame
# but you will lose the extensive header in the process
df = read.csv("harvard_DB_0001_1day.csv")
df = smooth_ts(df)
}
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
