% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.dates.r
\name{transition.dates}
\alias{transition.dates}
\title{Calculates transition dates on the upward (spring)
segments of a PhenoCam time series. This function
should not be run stand-alone. The required input
is a PhenoCam data file or data frame smoothed 
with the smooth.ts() routine using the BCI
optimized smoothing parameter}
\usage{
transition.dates(df, lower.thresh = 0.2, upper.thresh = 0.8,
  reverse = TRUE)
}
\arguments{
\item{df:}{a PhenoCam data file or data frame}

\item{lower.thresh:}{the minimum threshold used 
to determine the transition dates (default=0.2)}

\item{upper.thresh:}{the maximum threshold used
to determine the transition dates (default=0.8)}

\item{reverse:}{flip the direction of processing
if TRUE you calculate the senescence phase of the season}
}
\description{
Calculates transition dates on the upward (spring)
segments of a PhenoCam time series. This function
should not be run stand-alone. The required input
is a PhenoCam data file or data frame smoothed 
with the smooth.ts() routine using the BCI
optimized smoothing parameter
}
\examples{
# with defaults, outputting a data frame of transition dates
# for both the smoothed time series as well as the CI on this
# smoothed time series

my_dates <- transition.dates(df)

# dates need to be converted to standard notation using
as.Date(my_dates)
}
\keyword{PhenoCam,}
\keyword{dates,}
\keyword{phenology,}
\keyword{series}
\keyword{time}
\keyword{transition}

