% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_phenocam.r
\name{truncate_phenocam}
\alias{truncate_phenocam}
\title{Truncate a PhenoCam file using a given year as last valid year.
expand.phenocam() provides a similar functionality and is to be
prefered. This function remains as it might serve a purpose to
some.}
\usage{
truncate_phenocam(df, year = 2015)
}
\arguments{
\item{df}{a PhenoCam data frame}

\item{year}{the last valid year, discard the rest}
}
\description{
Truncate a PhenoCam file using a given year as last valid year.
expand.phenocam() provides a similar functionality and is to be
prefered. This function remains as it might serve a purpose to
some.
}
\examples{

\dontrun{
df = download_phenocam(site = "harvard",
                       type = "DB",
                       roi = "1",
                       frequency = 3)

truncate_phenocam("harvard_DB_0001_1day.csv", year = 2015)
# overwrites the original file, increasing
# decreasing the file size a given year as maximum.
}
}
\keyword{phenocam}
\keyword{series,}
\keyword{smoothing,}
\keyword{time}
